[4/21] Move code from app to controllers
DownloadController
